src/requirements.txt
azure-mgmt-resource
azure-mgmt-intune
msal
